Changed all ints to floats for ai tracking enemy to work

+++ src/com/NathanielEllsworth/main/BasicEnemy.java
@@ -26,7 +26,7 @@ public class BasicEnemy extends GameObject{
 	}
 	
 	public Rectangle getBounds(){
-		return new Rectangle(x, y, 16, 16); //rectangle collision bounds for basic enemy (loss of health)
+		return new Rectangle((int)x, (int)y, 16, 16); //rectangle collision bounds for basic enemy (loss of health)
 	}
 	
 
@@ -50,7 +50,7 @@ public class BasicEnemy extends GameObject{
 	public void render(Graphics g) {
 
 		g.setColor(Color.red);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x, (int)y, 16, 16);
 		
 	}
